Struct betrusted_pac::timer0::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {
    pub load3: LOAD3,
    pub load2: LOAD2,
    pub load1: LOAD1,
    pub load0: LOAD0,
    pub reload3: RELOAD3,
    pub reload2: RELOAD2,
    pub reload1: RELOAD1,
    pub reload0: RELOAD0,
    pub en: EN,
    pub update_value: UPDATE_VALUE,
    pub value3: VALUE3,
    pub value2: VALUE2,
    pub value1: VALUE1,
    pub value0: VALUE0,
    pub ev_status: EV_STATUS,
    pub ev_pending: EV_PENDING,
    pub ev_enable: EV_ENABLE,
}

Register block

Fields
load3: LOAD3

0x00 - Bits 24-31 of TIMER0_LOAD. Load value when Timer is (re-)enabled.In One-Shot mode, the value written to this register specify the Timer's duration in clock cycles.

load2: LOAD2

0x04 - Bits 16-23 of TIMER0_LOAD.

load1: LOAD1

0x08 - Bits 8-15 of TIMER0_LOAD.

load0: LOAD0

0x0c - Bits 0-7 of TIMER0_LOAD.

reload3: RELOAD3

0x10 - Bits 24-31 of TIMER0_RELOAD. Reload value when Timer reaches 0.In Periodic mode, the value written to this register specify the Timer's period in clock cycles.

reload2: RELOAD2

0x14 - Bits 16-23 of TIMER0_RELOAD.

reload1: RELOAD1

0x18 - Bits 8-15 of TIMER0_RELOAD.

reload0: RELOAD0

0x1c - Bits 0-7 of TIMER0_RELOAD.

en: EN

0x20 - Enable of the Timer.Set if to 1 to enable/start the Timer and 0 to disable the Timer

update_value: UPDATE_VALUE

0x24 - Update of the current countdown value.A write to this register latches the current countdown value to value register.

value3: VALUE3

0x28 - Bits 24-31 of TIMER0_VALUE. Latched countdown value

value2: VALUE2

0x2c - Bits 16-23 of TIMER0_VALUE.

value1: VALUE1

0x30 - Bits 8-15 of TIMER0_VALUE.

value0: VALUE0

0x34 - Bits 0-7 of TIMER0_VALUE.

ev_status: EV_STATUS

0x38 -

ev_pending: EV_PENDING

0x3c -

ev_enable: EV_ENABLE

0x40 -